# docs/api/openapi-admin.yaml
openapi: 3.1.0
info:
  title: RADIANT Admin API
  description: |
    Admin API for RADIANT multi-tenant AI platform.
    Provides endpoints for tenant management, AI model configuration, 
    billing, reports, and system administration.
  version: 5.42.0
  contact:
    name: RADIANT Support
  license:
    name: Proprietary
servers:
  - url: https://api.{domain}/admin
    description: Production
    variables:
      domain:
        default: radiant.ai
  - url: http://localhost:3001/admin
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /tenants:
    get:
      tags: [Tenants]
      summary: List all tenants
      operationId: listTenants
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, trial]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  total:
                    type: integer
    post:
      tags: [Tenants]
      summary: Create a new tenant
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /models:
    get:
      tags: [Models]
      summary: List available AI models
      operationId: listModels
      parameters:
        - name: provider
          in: query
          schema:
            type: string
        - name: capability
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/Model'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
  schemas:
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, trial]
        tier:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTenantRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        tier:
          type: integer
          default: 1
        settings:
          type: object
...
    Model:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        capabilities:
          type: array
          items:
            type: string
        context_window:
          type: integer
        cost_per_1k_input:
          type: number
        cost_per_1k_output:
          type: number
        is_enabled:
          type: boolean
...